refactor: Movies를 local db로 취급
